requirements.txt
aiosmtplib==2.0.1
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
DateTime==5.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.91.0
fastapi-mail==1.2.5
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
jose==1.0.0
MarkupSafe==2.1.2
outcome==1.2.0
passlib==1.7.4
Pillow==9.4.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pymongo==4.3.3
PySocks==1.7.1
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.24.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.14
uuid==1.30
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wsproto==1.2.0
zope.interface==5.5.2
